Example: Midi triggered samples played from flash (#154)
* Move common things from sample_player and flash_rw samples to core

* Small int type fixups in buffer_source.h and sound.h

* Fix some warnings in Mixer

* WIP: Add flash_audio_streaming example

* Move init_filesystem to core/filesystem.h

* WIP: things are broken

* WIP: pico_audio_i2s_32b causes write errors. Probably interrupts.

* Remove pico_audio_i2s_32b submodule

* Add pico_audio_i2s_32b as directory

* Remove pico_multicore dependency from pico_audio_i2s_32b, since it breaks flash writing

* Flash writing works while pico_audio_i2s_32b is linked

* file pointer in flash_rw example

* WIP: Write sample to flash and read the size

* Playing sample, but not from flash

* Hacky start, plays aliased noise

* Comment about source format conversion

* WIP: Playback from flash works with dumb temporary write buffer

* Playing snare sample from flash works!

* Plays sample while being a usb midi device

* Plays 4 samples, pitched by velocity, from flash

* Wrap notes

* Remove pico_audio_i2s_32b directory

* Add pico_audio_i2s_32b as submodule again

* Disable stdio-over-usb

* Disable warnings, since pico_audio_i2s_32b causes errors together with vfs

+++ ports/pico/core/sound.h
@@ -4,11 +4,11 @@
 #include "buffer_source.h"
 #include "timestretched/audio_memory_reader.h"
 #include "timestretched/pitch_shifter.h"
+#include <pico/stdlib.h>
 #include <stdint.h>
-// #include <stdio.h>
 
 struct Sound : BufferSource {
-  Sound(const unsigned int *sample_data, const size_t data_length)
+  Sound(const unsigned int *sample_data, const uint32_t data_length)
       : memory_reader(sample_data, data_length), pitch_shifter(memory_reader) {
   }
